+++ apps/dashboard/src/hooks/use-parse-variables.ts
@@ -1,4 +1,4 @@
-import { parseStepVariables } from '@/utils/parseStepVariables';
+import { parseStepVariables, type EnhancedParsedVariables } from '@/utils/parseStepVariables';
 import { type JSONSchemaDefinition } from '@novu/shared';
 import { JSONSchema7 } from 'json-schema';
 import { useMemo } from 'react';
@@ -7,7 +7,7 @@ export function useParseVariables(
   schema?: JSONSchemaDefinition | JSONSchema7,
   digestStepId?: string,
   isPayloadSchemaEnabled?: boolean
-) {
+): EnhancedParsedVariables {
   const parsedVariables = useMemo(() => {
     return schema
       ? parseStepVariables(schema, { digestStepId, isPayloadSchemaEnabled })
@@ -16,6 +16,7 @@ export function useParseVariables(
           namespaces: [],
           primitives: [],
           arrays: [],
+          enhancedVariables: [],
           isAllowedVariable: () => false,
         };
   }, [schema, digestStepId, isPayloadSchemaEnabled]);
