+++ apps/dashboard/src/utils/parseStepVariables.ts
@@ -9,12 +9,6 @@ import {
   getDynamicDigestVariable,
 } from '../components/variable/utils/digest-variables';
 import { JSONSchema7 } from 'json-schema';
-import {
-  mapJsonSchemaTypeToFieldType,
-  getInputTypeFromSchema,
-  type FieldDataType,
-  type EnhancedField,
-} from './schema-to-field-types';
 
 export interface LiquidVariable {
   type?: 'variable' | 'digest' | 'new-variable' | 'local';
@@ -25,6 +19,8 @@ export interface LiquidVariable {
   aliasFor?: string | null;
   isNewSuggestion?: boolean;
 }
+
+export type FieldDataType = 'string' | 'number' | 'boolean' | 'date' | 'datetime' | 'array' | 'object';
 
 export interface EnhancedLiquidVariable extends LiquidVariable {
   dataType: FieldDataType;
@@ -47,16 +43,53 @@ export interface EnhancedParsedVariables extends ParsedVariables {
   enhancedVariables: EnhancedLiquidVariable[];
 }
 
-/**
- * Parse JSON Schema and extract variables for Liquid autocompletion.
- * @param schema - The JSON Schema to parse.
- * @returns An object containing three arrays: primitives, arrays, and namespaces.
- */
+function mapJsonSchemaTypeToFieldType(schemaProperty: JSONSchemaDefinition | JSONSchema7): FieldDataType {
+  if (typeof schemaProperty === 'boolean') return 'string';
+
+  const { type, format } = schemaProperty;
+
+  switch (type) {
+    case 'string':
+      if (format === 'date') return 'date';
+      if (format === 'date-time') return 'datetime';
+      return 'string';
+    case 'number':
+    case 'integer':
+      return 'number';
+    case 'boolean':
+      return 'boolean';
+    case 'array':
+      return 'array';
+    case 'object':
+      return 'object';
+    default:
+      return 'string';
+  }
+}
+
+function getInputTypeFromSchema(schemaProperty: JSONSchemaDefinition | JSONSchema7): string {
+  if (typeof schemaProperty === 'boolean') return 'text';
+
+  const { type, format } = schemaProperty;
+
+  switch (type) {
+    case 'number':
+    case 'integer':
+      return 'number';
+    case 'string':
+      if (format === 'date') return 'date';
+      if (format === 'date-time') return 'datetime-local';
+      if (format === 'email') return 'email';
+      return 'text';
+    default:
+      return 'text';
+  }
+}
 
 export function parseStepVariables(
   schema: JSONSchemaDefinition | JSONSchema7,
   { digestStepId, isPayloadSchemaEnabled }: { digestStepId?: string; isPayloadSchemaEnabled?: boolean }
-): ParsedVariables {
+): EnhancedParsedVariables {
   const result: ParsedVariables = {
     primitives: [],
     arrays: [],
@@ -64,6 +97,8 @@ export function parseStepVariables(
     namespaces: [],
     isAllowedVariable: () => false,
   };
+
+  const enhancedVariables: EnhancedLiquidVariable[] = [];
 
   function extractProperties(obj: JSONSchemaDefinition | JSONSchema7, path = ''): void {
     if (typeof obj === 'boolean') return;
@@ -76,8 +111,10 @@ export function parseStepVariables(
 
         if (typeof value === 'object') {
           if (value.type === 'array') {
-            result.arrays.push({
+            result.arrays.push({ name: fullPath });
+            enhancedVariables.push({
               name: fullPath,
+              dataType: 'array',
             });
 
             if (value.properties) {
@@ -89,14 +126,23 @@ export function parseStepVariables(
               extractProperties(items, `${fullPath}[0]`);
             }
           } else if (value.type === 'object') {
-            if (fullPath !== 'payload' && fullPath !== 'steps') {
-              result.namespaces.push({ name: fullPath });
-            }
+            result.namespaces.push({ name: fullPath });
+            enhancedVariables.push({
+              name: fullPath,
+              dataType: 'object',
+            });
 
             extractProperties(value, fullPath);
           } else if (value.type && ['string', 'number', 'boolean', 'integer'].includes(value.type as string)) {
-            result.primitives.push({
+            const dataType = mapJsonSchemaTypeToFieldType(value);
+            const inputType = getInputTypeFromSchema(value);
+
+            result.primitives.push({ name: fullPath });
+            enhancedVariables.push({
               name: fullPath,
+              dataType,
+              inputType,
+              format: value.format,
             });
           }
         }
@@ -194,6 +240,26 @@ export function parseStepVariables(
     return true;
   }
 
+  // Add digest variables with type information
+  if (digestStepId) {
+    const digestVariables = DIGEST_VARIABLES.map((variable) => {
+      const { label: displayLabel, value } = getDynamicDigestVariable({
+        digestStepName: digestStepId,
+        type: variable.name as DIGEST_VARIABLES_ENUM,
+      });
+
+      return {
+        ...variable,
+        name: value,
+        displayLabel,
+        dataType: 'string' as FieldDataType,
+        inputType: 'text',
+      };
+    });
+
+    enhancedVariables.unshift(...digestVariables);
+  }
+
   return {
     ...result,
 
@@ -218,104 +284,6 @@ export function parseStepVariables(
       : [...result.primitives, ...result.arrays, ...result.namespaces],
 
     isAllowedVariable,
-  };
-}
-
-/**
- * Enhanced version of parseStepVariables that includes type information for React Query Builder
- */
-export function parseStepVariablesWithTypes(
-  schema: JSONSchemaDefinition | JSONSchema7,
-  { digestStepId, isPayloadSchemaEnabled }: { digestStepId?: string; isPayloadSchemaEnabled?: boolean }
-): EnhancedParsedVariables {
-  const baseResult = parseStepVariables(schema, { digestStepId, isPayloadSchemaEnabled });
-  const enhancedVariables: EnhancedLiquidVariable[] = [];
-
-  function extractPropertiesWithTypes(obj: JSONSchemaDefinition | JSONSchema7, path = ''): void {
-    if (typeof obj === 'boolean') return;
-
-    if (obj.type === 'object') {
-      if (!obj.properties) return;
-
-      for (const [key, value] of Object.entries(obj.properties)) {
-        const fullPath = path ? `${path}.${key}` : key;
-
-        if (typeof value === 'object') {
-          if (value.type === 'array') {
-            enhancedVariables.push({
-              name: fullPath,
-              dataType: 'array',
-            });
-
-            if (value.properties) {
-              extractPropertiesWithTypes({ type: 'object', properties: value.properties }, fullPath);
-            }
-
-            if (value.items) {
-              const items = Array.isArray(value.items) ? value.items[0] : value.items;
-              extractPropertiesWithTypes(items, `${fullPath}[0]`);
-            }
-          } else if (value.type === 'object') {
-            enhancedVariables.push({
-              name: fullPath,
-              dataType: 'object',
-            });
-
-            extractPropertiesWithTypes(value, fullPath);
-          } else if (value.type && ['string', 'number', 'boolean', 'integer'].includes(value.type as string)) {
-            const dataType = mapJsonSchemaTypeToFieldType(value);
-            const inputType = getInputTypeFromSchema(value);
-
-            enhancedVariables.push({
-              name: fullPath,
-              dataType,
-              inputType,
-              format: value.format,
-            });
-          }
-        }
-      }
-    }
-
-    // Handle combinators (allOf, anyOf, oneOf)
-    ['allOf', 'anyOf', 'oneOf'].forEach((combiner) => {
-      if (Array.isArray(obj[combiner as keyof typeof obj])) {
-        for (const subSchema of obj[combiner as keyof typeof obj] as JSONSchemaDefinition[]) {
-          extractPropertiesWithTypes(subSchema, path);
-        }
-      }
-    });
-
-    // Handle conditional schemas (if/then/else)
-    if (obj.if) extractPropertiesWithTypes(obj.if, path);
-    if (obj.then) extractPropertiesWithTypes(obj.then, path);
-    if (obj.else) extractPropertiesWithTypes(obj.else, path);
-  }
-
-  extractPropertiesWithTypes(schema);
-
-  // Add digest variables with default string type (only if digestStepId is provided)
-  if (digestStepId) {
-    const digestVariables = DIGEST_VARIABLES.map((variable) => {
-      const { label: displayLabel, value } = getDynamicDigestVariable({
-        digestStepName: digestStepId,
-        type: variable.name as DIGEST_VARIABLES_ENUM,
-      });
-
-      return {
-        ...variable,
-        name: value,
-        displayLabel,
-        dataType: 'string' as FieldDataType,
-        inputType: 'text',
-      };
-    });
-
-    enhancedVariables.unshift(...digestVariables);
-  }
-
-  return {
-    ...baseResult,
     enhancedVariables,
   };
 }
